refactor: move some stuff around
